odhcp6c: add failure when -E option is used without ubus support enabled
Problem: The -E option disables the state script and relies only on ubus to emit event on state changes.
But if odhcp6c is compiled without ubus support (ENABLE_UBUS flag omitted), odhcp6c has now way to signal state changes.

Solution: Return an error when -E flag is used and ubus support is disabled.

Issue: HOP-8608 - [CDROUTER][IPv6] The Box does not recover and sends SOLICIT when the DHCPv6 Server on WAN is not
Signed-off-by: Nicolas BESNARD <nicolas.besnard_ext@softathome.com>
GitOrigin-RevId: a511fc364edc34f7bc874b156f5c7d5d8e6c0f5c